Moving dns resolver logic to separate package
